requirements.txt.txt
discord.py==0.16.6
beautifulsoup4==4.6.0
